Remove all usages of aptly repositories
