--- mirror_mgmt/aptly/repository.py
@@ -1,48 +1,0 @@
-import subprocess
-
-from mirror_mgmt.exceptions import CallError
-
-from typing import Optional
-from typing_extensions import ParamSpec
-
-from .resource import Resource
-from .snapshot import Snapshot
-
-P = ParamSpec('P')
-
-
-class RepositoryBase(Resource):
-
-    def create_snapshot(self, snapshot_name: Optional[str] = None) -> Snapshot:
-        snap = Snapshot(
-            snapshot_name, self, distribution=self.distribution,
-            publish_prefix_override=self.publish_prefix_override,
-        )
-        if snap.exists:
-            snap.drop_published_snapshot()
-            snap.delete()
-
-        snap.create()
-        return snap
-
-    def drop(self):
-        self.run(['drop', '-force', self.resource_name])
-
-
-class Repository(RepositoryBase):
-
-    RESOURCE_NAME = 'repo'
-
-    def __init__(self, name: str, **kwargs: P.kwargs):
-        super().__init__(name, **kwargs)
-        self.package_directory = kwargs.get('package_directory')
-
-    def create(self) -> subprocess.CompletedProcess:
-        missing = [k for k in ('package_directory', 'distribution', 'name') if not getattr(self, k)]
-        if missing:
-            raise CallError(f'{", ".join(missing)!r} must be specified before attempting to create repository')
-
-        return self.run(['create', self.resource_name])
-
-    def update(self) -> None:
-        self.run(['add', '-remove-files', '-force-replace', self.resource_name, self.package_directory])
